+++ test/e2e/atlas/helper_test.go
@@ -27,6 +27,7 @@ import (
 
 	"github.com/mongodb/mongodb-atlas-cli/test/e2e"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	"go.mongodb.org/atlas/mongodbatlas"
 )
 
@@ -541,11 +542,30 @@ func createProjectWithoutAlertSettings(projectName string) (string, error) {
 	return project.ID, nil
 }
 
-func deleteAllNetworkPeers(projectID, provider string) error {
-	cliPath, err := e2e.AtlasCLIBin()
-	if err != nil {
-		return err
+func deleteClustersForProject(t *testing.T, cliPath, projectID string) {
+	t.Helper()
+	cmd := exec.Command(cliPath,
+		clustersEntity,
+		"list",
+		"--projectId", projectID,
+		"-o=json")
+	cmd.Env = os.Environ()
+	resp, err := cmd.CombinedOutput()
+	require.NoError(t, err, string(resp))
+	var clusters mongodbatlas.AdvancedClustersResponse
+	require.NoError(t, json.Unmarshal(resp, &clusters), string(resp))
+
+	for _, cluster := range clusters.Results {
+		clusterName := cluster.Name
+		t.Run("deleting cluster "+clusterName, func(t *testing.T) {
+			t.Parallel()
+			assert.NoError(t, deleteClusterForProject(projectID, clusterName))
+		})
 	}
+}
+
+func deleteAllNetworkPeers(t *testing.T, cliPath, projectID, provider string) {
+	t.Helper()
 	cmd := exec.Command(cliPath,
 		networkingEntity,
 		networkPeeringEntity,
@@ -558,38 +578,32 @@ func deleteAllNetworkPeers(projectID, provider string) error {
 	)
 	cmd.Env = os.Environ()
 	resp, err := cmd.CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("%s (%w)", string(resp), err)
-	}
+	require.NoError(t, err, string(resp))
 	var networkPeers []mongodbatlas.Peer
 	err = json.Unmarshal(resp, &networkPeers)
-	if err != nil {
-		return fmt.Errorf("invalid response: %s (%w). can't unmarshal", string(resp), err)
-	}
+	require.NoError(t, err)
 	for _, peer := range networkPeers {
-		cmd = exec.Command(cliPath,
-			networkingEntity,
-			networkPeeringEntity,
-			"delete",
-			peer.ID,
-			"--projectId",
-			projectID,
-			"--force",
-		)
-		cmd.Env = os.Environ()
-		resp, err = cmd.CombinedOutput()
-		if err != nil {
-			return fmt.Errorf("%s (%w)", string(resp), err)
-		}
+		peerID := peer.ID
+		t.Run("deleting "+peerID, func(t *testing.T) {
+			t.Parallel()
+			cmd = exec.Command(cliPath,
+				networkingEntity,
+				networkPeeringEntity,
+				"delete",
+				peerID,
+				"--projectId",
+				projectID,
+				"--force",
+			)
+			cmd.Env = os.Environ()
+			resp, err = cmd.CombinedOutput()
+			assert.NoError(t, err, string(resp))
+		})
 	}
-	return nil
 }
 
-func deleteAllPrivateEndpoints(projectID, provider string) error {
-	cliPath, err := e2e.AtlasCLIBin()
-	if err != nil {
-		return err
-	}
+func deleteAllPrivateEndpoints(t *testing.T, cliPath, projectID, provider string) {
+	t.Helper()
 	cmd := exec.Command(cliPath,
 		privateEndpointsEntity,
 		provider,
@@ -600,31 +614,28 @@ func deleteAllPrivateEndpoints(projectID, provider string) error {
 	)
 	cmd.Env = os.Environ()
 	resp, err := cmd.CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("%s (%w)", string(resp), err)
-	}
+	require.NoError(t, err, string(resp))
 	var privateEndpoints []mongodbatlas.PrivateEndpointConnection
 	err = json.Unmarshal(resp, &privateEndpoints)
-	if err != nil {
-		return fmt.Errorf("invalid response: %s (%w). can't unmarshal", string(resp), err)
-	}
+	require.NoError(t, err, string(resp))
 	for _, endpoint := range privateEndpoints {
-		cmd = exec.Command(cliPath,
-			privateEndpointsEntity,
-			provider,
-			"delete",
-			endpoint.ID,
-			"--projectId",
-			projectID,
-			"--force",
-		)
-		cmd.Env = os.Environ()
-		resp, err = cmd.CombinedOutput()
-		if err != nil {
-			return fmt.Errorf("%s (%w)", string(resp), err)
-		}
+		endpointID := endpoint.ID
+		t.Run("deleting "+endpointID, func(t *testing.T) {
+			t.Parallel()
+			cmd = exec.Command(cliPath,
+				privateEndpointsEntity,
+				provider,
+				"delete",
+				endpointID,
+				"--projectId",
+				projectID,
+				"--force",
+			)
+			cmd.Env = os.Environ()
+			resp, err = cmd.CombinedOutput()
+			assert.NoError(t, err, string(resp))
+		})
 	}
-	return nil
 }
 
 func deleteTeam(teamID string) error {
