+++ test/e2e/atlas/kubernetes_config_generate_test.go
@@ -636,7 +636,7 @@ func TestProjectWithNetworkPeering(t *testing.T) {
 		resp, err := cmd.CombinedOutput()
 		require.NoError(t, err)
 		t.Cleanup(func() {
-			deleteNetworkPeering(t, generator.projectID, gcpEntity)
+			deleteAllNetworkPeers(t, cliPath, generator.projectID, gcpEntity)
 		})
 		var createdNetworkPeer mongodbatlas.Peer
 		err = json.Unmarshal(resp, &createdNetworkPeer)
@@ -697,7 +697,7 @@ func TestProjectWithPrivateEndpoint_Azure(t *testing.T) {
 		resp, err := cmd.CombinedOutput()
 		require.NoError(t, err)
 		s.t.Cleanup(func() {
-			deletePrivateEndpoints(t, generator.projectID, azureEntity)
+			deleteAllPrivateEndpoints(t, cliPath, generator.projectID, azureEntity)
 		})
 		var createdNetworkPeer mongodbatlas.PrivateEndpointConnection
 		err = json.Unmarshal(resp, &createdNetworkPeer)
